Dev: move QuestionAttributesSettings to an helper Dev: add /ls/helper to namespace (think we need it elsewhere) Dev: now, we load only attribute we need (not whole question type attribute) Dev: in fact : we load all existing attribute, but take only needed for $questionAttributesSettings[type]
