+++ application/controllers/admin/questions.php
@@ -1361,7 +1361,7 @@ class questions extends Survey_Common_Action
              * Since is moved via ajax call only : it's not needed, when we have time : readd it for no-js solution
              */
             //~ if (!$adding)
-                //~ $qattributes = Question::getQuestionAttributesSettings($aqresult->type); //(or Question::getAdvancedSettingsWithValues )
+                //~ $qattributes = \ls\helpers\questionHelper::getQuestionAttributesSettings(($aqresult->type); //(or Question::getAdvancedSettingsWithValues )
             //~ else
                 //~ $qattributes = array();
 
@@ -1721,14 +1721,15 @@ class questions extends Survey_Common_Action
     */
     public function ajaxquestionattributes()
     {
+
         $surveyid = (int) Yii::app()->request->getParam('sid',0);
         $qid = (int) Yii::app()->request->getParam('qid',0);
         $type = Yii::app()->request->getParam('question_type');
         $thissurvey = getSurveyInfo($surveyid);
         if(!$thissurvey) die();
         $aLanguages = array_merge(array(Survey::model()->findByPk($surveyid)->language), Survey::model()->findByPk($surveyid)->additionalLanguages);
-
         $aAttributesWithValues = Question::model()->getAdvancedSettingsWithValues($qid, $type, $surveyid);
+
         uasort($aAttributesWithValues, 'categorySort');
 
         $aAttributesPrepared = array();
